apps/nlp/pyproject.toml
[project]
name = "valliance-graph-nlp"
version = "0.1.0"
description = "NLP enrichment pipeline for the Valliance Graph"
requires-python = ">=3.10"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
